refactor: estrae source_check_analyze.py da bulk_source_check.py
- Estratte funzioni di analisi pura in source_check_analyze.py
  (infer granularità/anni, parse CKAN, scoring, format)
- bulk_source_check.py passa da ~580 a ~300 LOC
- 100% test passano
