+++ scripts/source_check_analyze.py
@@ -1,0 +1,232 @@
+"""
+Analyze fase per bulk source-check.
+
+Funzioni di analisi pura (nessuna dipendenza HTTP o I/O).
+Estratto da bulk_source_check.py.
+"""
+from __future__ import annotations
+
+import re
+from typing import Any, Optional
+
+import pandas as pd
+
+from source_check_fetch import _EMPTY_ENRICH, _fetch_sdmx_years
+
+# ── euristica granularità ─────────────────────────────────────────────────────
+
+
+_GRAN_PATTERNS: list[tuple[str, str]] = [
+    (r"\bcomun[ei]\b|\bmunicip", "comune"),
+    (r"\bprovinc", "provincia"),
+    (
+        r"\bregion[ei]\b|\bregioni\b|piemonte|lombardia|veneto|emilia|toscana|lazio|campania|puglia|sicilia|sardegna|abruzzo|umbria|marche|molise|calabria|basilicata|friuli|trentin|liguria|valle d['\s]aosta",
+        "regione",
+    ),
+    (r"\bnazional[ei]\b|\bitali[ae]\b|\bnazione\b|\bnational\b|\bregional\b", "nazionale"),
+    (r"(?<![a-zA-Z])(REG|reg)(?![a-zA-Z])", "regione"),
+    (r"\beurope[ao]\b|\bue\b|\beuropa\b|\beuropean\b", "europeo"),
+]
+
+
+def _infer_granularity(text: str) -> str:
+    low = text.lower()
+    for pattern, label in _GRAN_PATTERNS:
+        if re.search(pattern, low):
+            return label
+    return "non_determinato"
+
+
+# ── euristica anni ────────────────────────────────────────────────────────────
+
+_YEAR_RE = re.compile(r"(?<!\d)(19\d{2}|20[012]\d)(?!\d)")
+_YEAR_START_RE = re.compile(r"(?:^|(?<!\d))(20[012]\d)")
+_COMPACT_YEAR_PAIR_RE = re.compile(r"(20[012]\d)(\d{2})(?=20|$|\D)")
+
+
+def _infer_years(text: str) -> tuple[Optional[int], Optional[int]]:
+    years: set[int] = set()
+    for y in _YEAR_RE.findall(text):
+        years.add(int(y))
+    for y in _YEAR_START_RE.findall(text):
+        years.add(int(y))
+    for first_str, second_str in _COMPACT_YEAR_PAIR_RE.findall(text):
+        y1 = int(first_str)
+        y2_2digit = int(second_str)
+        if y2_2digit <= 30:
+            y2 = 2000 + y2_2digit
+            if y2 > y1 and y2 - y1 <= 10:
+                years.add(y1)
+                years.add(y2)
+    if not years:
+        return None, None
+    return min(years), max(years)
+
+
+# ── CKAN analysis ─────────────────────────────────────────────────────────────
+
+
+def _parse_ckan_package(pkg: dict) -> dict:
+    """Estrae i campi utili da un package CKAN."""
+    tags = [
+        (t.get("display_name") or t.get("name") or "")
+        for t in (pkg.get("tags") or [])
+        if isinstance(t, dict)
+    ]
+
+    groups = [
+        (g.get("display_name") or g.get("name") or "")
+        for g in (pkg.get("groups") or [])
+        if isinstance(g, dict)
+    ]
+
+    resources = pkg.get("resources") or []
+    resource_url = None
+    resource_format = None
+    for res in resources:
+        u = res.get("url") or ""
+        if u.startswith("http"):
+            resource_url = u
+            resource_format = res.get("format") or None
+            break
+
+    extras = {e["key"]: e["value"] for e in (pkg.get("extras") or []) if isinstance(e, dict)}
+    temporal_start = extras.get("temporal_coverage_from") or extras.get("issued")
+    temporal_end = extras.get("temporal_coverage_to") or extras.get("modified")
+
+    if temporal_start is None and temporal_end is None:
+        periodo = extras.get("Periodo di riferimento") or extras.get("periodo di riferimento")
+        if periodo:
+            years = _YEAR_RE.findall(str(periodo))
+            if len(years) >= 2:
+                temporal_start, temporal_end = years[0], years[-1]
+
+    notes = (pkg.get("notes") or "").strip()
+    title = pkg.get("title") or None
+
+    combined = " ".join(filter(None, [title, ", ".join(groups), ", ".join(tags), notes[:500]]))
+    granularity = _infer_granularity(combined)
+
+    year_min, year_max = None, None
+    if temporal_start:
+        ys, _ = _infer_years(temporal_start)
+        year_min = ys
+    if temporal_end:
+        _, ye = _infer_years(temporal_end)
+        year_max = ye
+    if year_min is None or year_max is None:
+        yt_min, yt_max = _infer_years(combined)
+        year_min = year_min or yt_min
+        year_max = year_max or yt_max
+
+    return {
+        "enriched_title": title,
+        "enriched_tags": ", ".join(tags) if tags else None,
+        "enriched_notes": notes[:300] if notes else None,
+        "resource_url": resource_url,
+        "resource_format": resource_format,
+        "granularity": granularity,
+        "year_min": year_min,
+        "year_max": year_max,
+        "enrich_method": "ckan_package_show",
+    }
+
+
+# ── fallback euristica ────────────────────────────────────────────────────────
+
+
+def _fallback_infer(row: pd.Series) -> tuple[str, Optional[int], Optional[int]]:
+    parts = []
+    for col in ("title", "tags", "notes_excerpt", "prefix", "url"):
+        v = row.get(col)
+        if v and str(v) not in ("nan", "None", ""):
+            parts.append(str(v))
+    combined = " ".join(parts)
+    return _infer_granularity(combined), *_infer_years(combined)
+
+
+# ── intake scoring ────────────────────────────────────────────────────────────
+
+
+_GRAN_SCORE = {"comune": 40, "provincia": 30, "regione": 20, "nazionale": 10, "europeo": 5, "non_determinato": 0}
+_FORMAT_SCORE = {"CSV": 20, "JSON": 20, "XLSX": 12, "XLS": 10, "XML": 8, "SDMX": 8, "PDF": 2}
+_YEAR_SPAN_MAX = 20
+
+
+def _normalize_format(raw: str) -> str:
+    if not isinstance(raw, str):
+        return ""
+    valid_priority = ["CSV", "JSON", "XLSX", "XLS", "XML", "PDF", "SDMX", "ZIP", "PARQUET"]
+    up = raw.upper()
+    for v in valid_priority:
+        if v in up:
+            return v
+    return ""
+
+
+def _intake_score(
+    granularity: Optional[str],
+    year_min: Optional[int],
+    year_max: Optional[int],
+    reachable: bool,
+    resource_format: Optional[str],
+    enrich_method: str,
+    needs_review: bool,
+    source_status: Optional[str] = None,
+) -> tuple[int, bool]:
+    score = 0
+    score += _GRAN_SCORE.get(granularity or "non_determinato", 0)
+
+    if year_min is not None and year_max is not None:
+        span = max(0, year_max - year_min)
+        score += min(20, int(span / _YEAR_SPAN_MAX * 20))
+    elif year_min is not None or year_max is not None:
+        score += 5
+
+    score += 20 if reachable else 0
+
+    fmt_raw = ("" if not isinstance(resource_format, str) else resource_format).strip()
+    fmt = ""
+    if fmt_raw:
+        valid = ["CSV", "JSON", "XLSX", "XLS", "XML", "PDF", "SDMX", "ZIP", "PARQUET"]
+        if "." in fmt_raw and len(fmt_raw) > 6:
+            fmt_ext = fmt_raw.rsplit(".", 1)[-1].upper()
+            if fmt_ext in valid:
+                fmt = fmt_ext
+        else:
+            up = fmt_raw.upper()
+            for v in valid:
+                if v in up:
+                    fmt = v
+                    break
+    score += _FORMAT_SCORE.get(fmt, 0)
+
+    enrich_str = enrich_method if isinstance(enrich_method, str) else ""
+    if enrich_str in ("ckan_package_show", "sdmx_dataflow_annotations"):
+        score += 5
+    if needs_review:
+        score -= 5
+
+    if source_status == "stale":
+        score -= 10
+        needs_review = True
+
+    score = max(0, min(100, score))
+    candidate = score >= 40 and not needs_review
+    return score, candidate
+
+
+def _finalize_scores(result: dict) -> dict:
+    score, candidate = _intake_score(
+        granularity=result.get("granularity"),
+        year_min=result.get("year_min"),
+        year_max=result.get("year_max"),
+        reachable=result.get("reachable", False),
+        resource_format=result.get("resource_format"),
+        enrich_method=result.get("enrich_method", "none"),
+        needs_review=result.get("needs_review", True),
+        source_status=result.get("source_status"),
+    )
+    result["intake_score"] = score
+    result["intake_candidate"] = candidate
+    return result
